refactor: Moove config to root folder
