+++ integration-libs/cdc/public_api.ts
@@ -5,4 +5,4 @@ export * from './src/auth/facade/index';
 export * from './src/auth/store/actions/index';
 export * from './src/cdc.module';
 export * from './src/cms-components/gigya-raas/index';
-export * from './src/config/index';
+export * from './root/config/index';
