+++ integration-libs/cdc/root/config/cdc-config.ts
