+++ integration-libs/cdc/root/config/index.ts
