+++ integration-libs/cdc/src/auth/facade/cdc-js.service.ts
@@ -17,7 +17,7 @@ import {
 } from '@spartacus/core';
 import { combineLatest, Observable, ReplaySubject, Subscription } from 'rxjs';
 import { take } from 'rxjs/operators';
-import { CdcConfig } from '../../config/cdc-config';
+import { CdcConfig } from '../../../root/config/cdc-config';
 import { CdcAuthService } from './cdc-auth.service';
 
 @Injectable({
